restructure componenets and add page changing functionality
