+++ frontend/package.json
@@ -15,7 +15,8 @@
     "@fontsource/roboto": "^5.0.13",
     "@mui/material": "^5.15.20",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
